+++ src/flytetest/types/__init__.py
@@ -6,6 +6,7 @@
 
 from flytetest.types.assets import (
     AbInitioResultBundle,
+    AnnotationRefinementResultBundle,
     AssetToolProvenance,
     Braker3InputBundleAsset,
     Braker3NormalizedGff3Asset,
@@ -16,6 +17,7 @@ from flytetest.types.assets import (
     ChunkedProteinFastaAsset,
     EvmBraker3InputBundle,
     EvmCommandSetAsset,
+    ConsensusAnnotationResultBundle,
     EvmConsensusResultBundle,
     EvmExecutionInputBundleAsset,
     EvmInputPreparationBundle,
@@ -28,6 +30,7 @@ from flytetest.types.assets import (
     EvmTranscriptInputBundleAsset,
     EvmProteinEvidenceGff3Asset,
     ExonerateChunkAlignmentResult,
+    ProteinAlignmentChunkResult,
     ManifestSerializable,
     MergedBamAsset,
     PasaAlignmentAssemblyResult,
@@ -47,6 +50,7 @@ from flytetest.types.assets import (
     StarAlignmentResult,
     StarGenomeIndexAsset,
     StringTieAssemblyResult,
+    CodingPredictionResult,
     TransDecoderPredictionResult,
     TrinityDeNovoTranscriptAsset,
     TrinityGenomeGuidedAssemblyResult,
@@ -55,6 +59,7 @@ from flytetest.types.assets import (
 
 __all__ = [
     "AbInitioResultBundle",
+    "AnnotationRefinementResultBundle",
     "AssetToolProvenance",
     "Braker3InputBundleAsset",
     "Braker3NormalizedGff3Asset",
@@ -62,6 +67,7 @@ __all__ = [
     "Braker3ResultBundle",
     "CleanedTranscriptDataset",
     "EvmBraker3InputBundle",
+    "ConsensusAnnotationResultBundle",
     "EvmCommandSetAsset",
     "EvmConsensusResultBundle",
     "EvmExecutionInputBundleAsset",
@@ -77,6 +83,7 @@ __all__ = [
     "ChunkedProteinFastaAsset",
     "EvmProteinEvidenceGff3Asset",
     "ExonerateChunkAlignmentResult",
+    "ProteinAlignmentChunkResult",
     "ManifestSerializable",
     "MergedBamAsset",
     "PasaAlignmentAssemblyResult",
@@ -96,6 +103,7 @@ __all__ = [
     "StarAlignmentResult",
     "StarGenomeIndexAsset",
     "StringTieAssemblyResult",
+    "CodingPredictionResult",
     "TransDecoderPredictionResult",
     "TrinityDeNovoTranscriptAsset",
     "TrinityGenomeGuidedAssemblyResult",